@@ -10,7 +10,7 @@
 
     "config": {
         "platform": {
-            "php": "5.5.38"
+            "php": "7.3.0"
         }
     },
     "require": {
@@ -22,7 +22,7 @@
     },
     "require-dev": {
         "symfony/filesystem": "~2.0|~3.0|~4.0|~5.0",
-        "phpunit/phpunit": "^4.8.0",
+        "phpunit/phpunit": "^9.0.0",
         "bamarni/composer-bin-plugin": "^1.2"
     },
     "license" : "Apache-2.0"
