@@ -7,7 +7,7 @@ use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Process\Exception\ProcessFailedException;
 use Symfony\Component\Process\Process;
 
-class RepositoryInspectorTest extends \PHPUnit_Framework_TestCase
+class RepositoryInspectorTest extends \PHPUnit\Framework\TestCase
 {
     private $tmpDirs = array();
 
@@ -37,19 +37,19 @@ class RepositoryInspectorTest extends \PHPUnit_Framework_TestCase
         $tmpDir = $this->getTempDir();
         mkdir($tmpDir, 0777, true);
 
-        $this->exec('git init', $tmpDir);
+        $this->exec(['git', 'init'], $tmpDir);
         file_put_contents($tmpDir.'/foo', 'foo');
-        $this->exec('git add . && git commit -m "adds foo"', $tmpDir);
+        $this->exec(['sh', '-c', 'git add . && git commit -m "adds foo"'], $tmpDir);
 
         $introspector = new RepositoryIntrospector($tmpDir);
         $headRev = $introspector->getCurrentRevision();
 
         file_put_contents($tmpDir.'/bar', 'bar');
-        $this->exec('git add . && git commit -m "adds bar"', $tmpDir);
+        $this->exec(['sh', '-c', 'git add . && git commit -m "adds bar"'], $tmpDir);
         $this->assertEquals(array($headRev), $introspector->getCurrentParents());
     }
 
-    protected function tearDown()
+    protected function tearDown(): void
     {
         parent::tearDown();
 
@@ -77,7 +77,7 @@ class RepositoryInspectorTest extends \PHPUnit_Framework_TestCase
 
     private function installRepository($url, $dir)
     {
-        $proc = new Process('git clone '.$url.' '.$dir);
+        $proc = new Process(['git', 'clone', $url, $dir]);
         if (0 !== $proc->run()) {
             throw new ProcessFailedException($proc);
         }
